Additional information, fixed typos and examples.
Thanks for the library, hope to contribute!

After spending some time with support trying to figure out what the correct URL is, why the logs weren't received, and why exceptions didn't bubble up, I had a look at the source code and figured out why. Hopefully this will make it easier for the next developer to get started.

BTW, the documentation on the DataDog website is not up to date, and the wrong URL is used, and what I assume is earlier versions of this library.