Remove Docs Team as CODEOWNERS (#140)
Removes Docs as CODEOWNERS from the repo. As part of a project to reduce
courtesy reviews, the Docs Team is limiting CODEOWNERS entries to repos
that single-source content into the docs site or where our review is
critical. We’re still happy to review content if you reach out to us
directly.